MOD #157 : Product 구조 리팩토링
